requirement.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.7
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-import-export==0.6.1
django-reversion==2.0.12
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
dlib==19.8.1
et-xmlfile==1.0.1
face-recognition==1.0.0
face-recognition-models==0.3.0
future==0.16.0
httplib2==0.10.3
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.0
odfpy==1.3.6
opencv-python==3.4.0.12
openpyxl==2.4.9
Pillow==5.0.0
PyJWT==1.5.3
pytz==2017.3
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
